# openapi/spec.yaml
openapi: 3.0.0
info:
  title: SupALLM API
  version: "0.0.1"

# Apply security globally
security:
  - BearerAuth: []
paths:
  /projects:
    get:
      summary: List all projects
      operationId: listProjects
      tags:
        - Project
      responses:
        "200":
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
    post:
      summary: "Create a project"
      operationId: createProject
      tags:
        - Project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectRequest"
      responses:
        "201":
          description: "Project created"
  /projects/{projectId}:
    get:
      summary: "Get a project by ID"
      operationId: getProject
      tags:
        - Project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/UUID"
      responses:
        "200":
          description: "Project details"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "404":
          description: "Project not found"
    put:
      summary: "Update a project"
      operationId: updateProject
      tags:
        - Project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/UUID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProjectRequest"
      responses:
        "200":
          description: "Project updated"
        "404":
          description: "Project not found"
    delete:
      summary: "Delete a project"
      operationId: deleteProject
      tags:
        - Project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/UUID"
      responses:
        "204":
          description: "Project deleted"
        "404":
          description: "Project not found"
  /projects/{projectId}/auth:
    put:
      summary: "Update authentication configuration for a project"
      operationId: updateAuth
      tags:
        - Project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/UUID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAuthRequest"
      responses:
        "200":
          description: "Authentication configuration updated"
        "404":
          description: "Project not found"
...
  /projects/{projectId}/models:
    get:
      summary: List all models for a project
      operationId: listModels
      tags:
        - Model
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/UUID"
      responses:
        "200":
          description: List of models
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Model"
        "404":
          description: Project not found
    post:
      summary: "Create a model for a project"
      operationId: createModel
      tags:
        - Model
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/UUID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateModelRequest"
      responses:
        "201":
          description: "Model created"
        "400":
          description: Bad request
        "404":
          description: Project not found
  /projects/{projectId}/models/{slug}:
    get:
      summary: "Get a model by slug"
      operationId: getModel
      tags:
        - Model
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/UUID"
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: "Model details"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Model"
        "404":
          description: Model or project not found
    put:
      summary: "Update a model"
      operationId: updateModel
      tags:
        - Model
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/UUID"
        - name: slug
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateModelRequest"
      responses:
        "200":
          description: "Model updated"
        "400":
          description: Bad request
        "404":
          description: Model or project not found
    delete:
      summary: "Delete a model"
      operationId: deleteModel
      tags:
        - Model
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/UUID"
        - name: slug
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: "Model deleted"
        "404":
          description: Model or project not found
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    UUID:
      type: string
      x-go-type: uuid.UUID
      x-go-type-import:
        name: uuid
        path: github.com/google/uuid
    Project:
      type: object
      properties:
        id:
          $ref: "#/components/schemas/UUID"
        name:
          type: string
        authProvider:
          $ref: "#/components/schemas/AuthProvider"
        credentials:
          type: array
          items:
            $ref: "#/components/schemas/Credential"
        models:
          type: array
          items:
            $ref: "#/components/schemas/Model"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - userId
        - authProvider
        - credentials
        - models
        - createdAt
        - updatedAt
    CreateProjectRequest:
      type: object
      properties:
        name:
          type: string
      required:
        - name
    UpdateProjectRequest:
      type: object
      properties:
        name:
          type: string
      required:
        - name
    ProviderType:
      type: string
      enum:
        - openai
        - anthropic
    ProviderModel:
      type: string
      enum:
        - gpt-4o
        - gpt-4o-mini
        - claude-3-7-sonnet
        - claude-3-5-sonnet
        - claude-3-5-haiku
    ModelParameters:
      type: object
      properties:
        maxTokens:
          type: integer
        temperature:
          type: number
      required:
        - maxTokens
        - temperature
    Credential:
      type: object
      properties:
        id:
          $ref: "#/components/schemas/UUID"
        name:
          type: string
        provider:
          $ref: "#/components/schemas/ProviderType"
        apiKey:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - provider
        - apiKey
        - createdAt
        - updatedAt
...
    Model:
      type: object
      properties:
        slug:
          type: string
        name:
          type: string
        credentialId:
          $ref: "#/components/schemas/UUID"
        providerModel:
          $ref: "#/components/schemas/ProviderModel"
        systemPrompt:
          type: string
        parameters:
          $ref: "#/components/schemas/ModelParameters"
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - slug
        - credentialId
        - providerModel
        - systemPrompt
        - parameters
        - createdAt
        - updatedAt
    CreateModelRequest:
      type: object
      properties:
        name:
          type: string
        credentialId:
          type: string
          format: uuid
        providerModel:
          type: string
        systemPrompt:
          type: string
        parameters:
          $ref: "#/components/schemas/ModelParameters"
      required:
        - name
        - credentialId
        - providerModel
        - systemPrompt
        - parameters
    UpdateModelRequest:
      type: object
      properties:
        name:
          type: string
        credentialId:
          type: string
          format: uuid
        providerModel:
          $ref: "#/components/schemas/ProviderModel"
        systemPrompt:
          type: string
        parameters:
          $ref: "#/components/schemas/ModelParameters"
      required:
        - name
        - credentialId
        - providerModel
        - systemPrompt
        - parameters
    AuthProvider:
      type: object
      properties:
        provider:
          type: string
          enum:
            - supabase
            - clerk
            - firebase
        config:
          type: object
      required:
        - provider
        - config
    UpdateAuthRequest:
      type: object
      properties:
        provider:
          type: string
          enum:
            - supabase
            - clerk
            - firebase
        config:
          type: object
      required:
        - provider
        - config
...